requirements.txt
edge-tts
moviepy
Pillow
requests
python-dotenv
bytez
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
huggingface_hub
tzdata
